requirements.txt
cryptography==44.0.2
flask==3.0.2
flask-restx==1.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
